refactor: tidy up before build the main course app
