+++ components/exercises/fetch-data/fragments/Loading.tsx
