+++ components/exercises/initial-exercises/ShowMore.tsx
