+++ components/exercises/toggle-colors/ManualToggle.tsx
