+++ package.json
@@ -26,7 +26,6 @@
 		"@types/node": "20.5.9",
 		"@types/react": "18.2.21",
 		"@types/react-dom": "18.2.7",
-		"@uidotdev/usehooks": "^2.3.1",
 		"@vercel/analytics": "^1.0.2",
 		"autoprefixer": "10.4.15",
 		"axios": "^1.5.0",
@@ -36,6 +35,7 @@
 		"eslint-config-next": "13.4.19",
 		"lucide-react": "^0.274.0",
 		"next": "13.4.19",
+		"next-themes": "^0.2.1",
 		"postcss": "8.4.29",
 		"react": "18.2.0",
 		"react-dom": "18.2.0",
